requirements.txt
torch>=1.12.0
transformers>=4.20.0
datasets>=2.0.0
mlflow>=2.0.0
fastapi>=0.85.0
uvicorn[standard]>=0.18.0
dvc[s3]>=2.0.0
apache-airflow>=2.5.0
pydantic
